added help check and 3rd variable to work with deployProcessor correctly

+++ final_parts/Tom's_folder/installBundle.sh
@@ -1,8 +1,17 @@
 #!/usr/bin/bash
+set -e
 
 bundlePath="$1"
 targetPath="$2"
+serviceName="$3"
+
+if [ "$bundlePath" = "h" ] || [ "$bundlePath" = "-h" ] || [ "$bundlePath" = "help" ] || [ "$bundlePath" = "-help" ]; then
+    echo "./installBundle.sh <bundlePath> <targetPath>"
+    exit
+fi
 
 mkdir -p "$targetPath"
 rm -rf "${targetPath:?}/"*
 tar -xf "$bundlePath" -C "$targetPath"
+
+systemctl restart "$serviceName"
